Fix invalid pins

+++ port/portserial.c
@@ -61,7 +61,7 @@
 /* ----------------------- Static variables ---------------------------------*/
 
 typedef struct {
-    uint8_t pin;
+    uint16_t pin;
     GPIO_TypeDef *gpio;
     GPIOSpeed_TypeDef speed;
     GPIOMode_TypeDef mode;
